pyproject.toml
[tool.poetry]
name = "flask-crud"
version = "0.1.0"
description = ""
authors = ["Fabio Araujo <fabioaraujods@gmail.com>"]
readme = "README.md"
packages = [{include = "flask_crud"}]

[tool.poetry.dependencies]
python = "^3.9.9"
Flask = "^2.0.2"
Flask-SQLAlchemy = "3.0.3"
Flask-WTF = "1.0.0"
Flask-Bcrypt = "0.7.1"
mysql = "^0.0.3"
mysql-connector-python = "^8.0.33"
python-dotenv = "^1.0.0"



[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
